requirements.txt
requests
pandas
python-dotenv
matplotlib
streamlit>=1.31
altair>=5
pytest
pytest-cov
ruff
pandera
plotly
pandas-gbq
google-auth
google-cloud-bigquery
numpy
gridstatus>=0.30
openpyxl>=3.1